moved config file and commented on rope scaling.
