+++ src/instructlab/dolomite/hf_models/__init__.py
@@ -2,8 +2,9 @@
 # Extracted from https://github.com/ibm-granite/dolomite-engine
 # ----------------------------------------------------------------
 # Local
+from .config import GPTDolomiteConfig
 from .model_conversion import export_to_huggingface, import_from_huggingface
-from .models import GPTDolomiteConfig, GPTDolomiteForCausalLM, GPTDolomiteModel
+from .models import GPTDolomiteForCausalLM, GPTDolomiteModel
 from .register_hf import register_model_classes
 
 register_model_classes()
