+++ src/instructlab/dolomite/hf_models/modeling_utils/attention/__init__.py
@@ -9,8 +9,8 @@ import inspect
 import torch
 
 # Local
+from ...config import GPTDolomiteConfig
 from ...enums import AttentionHeadType
-from ...models.gpt_dolomite.config import GPTDolomiteConfig
 from .base import Attention
 from .flash import FlashAttention2
 from .padding_free import PaddingFreeAttention
