+++ src/instructlab/dolomite/hf_models/models/gpt_dolomite/__init__.py
@@ -3,6 +3,5 @@
 # ----------------------------------------------------------------
 # Local
 from .base import GPTDolomiteModel, GPTDolomitePreTrainedModel
-from .config import GPTDolomiteConfig
 from .main import GPTDolomiteForCausalLM
 from .mlp import interleave_up_gate_tensor_for_mlp, split_up_gate_tensor_for_mlp
